Even Or Odd Program
In class exercise that tells if a number input even or odd using a ternary operator. The logic for this is represented that an even number can be represented with the formula n = 2k where k is some int and any odd number can be represented with n = 2k + 1 where k is also some int.