Hash Table Program
Creates a more hash function, though it is nothing like a SHA-1 hash or anything it is a little more advanced. Also, it uses this algorithm to create a hash table that revolves collisions with chaining. After this it allows the user to search the hash table and such, much more scalable that other searching algorithms as it can approach O(1).